Extract async sibling candidate resolution
